refactor: move CreateModulePayload to createModule.ts
